qtools/requirements.txt
python==3.9.7
tensorflow==2.4.1
biopython
pandas
scikit-learn
scipy 
seaborn
numpy==1.21.2
logomaker
qtools
